+++ app/controllers/shops_controller.rb
@@ -1,70 +1,11 @@
-class ShopsController < ApplicationController
-  before_action :set_shop, only: %i[ show edit update destroy ]
+# app/controllers/shops_controller.rb
 
-  # GET /shops or /shops.json
+class ShopsController < ApplicationController
   def index
     @shops = Shop.all
   end
 
-  # GET /shops/1 or /shops/1.json
   def show
+    @shop = Shop.find(params[:id])
   end
-
-  # GET /shops/new
-  def new
-    @shop = Shop.new
-  end
-
-  # GET /shops/1/edit
-  def edit
-  end
-
-  # POST /shops or /shops.json
-  def create
-    @shop = Shop.new(shop_params)
-
-    respond_to do |format|
-      if @shop.save
-        format.html { redirect_to shop_url(@shop), notice: "Shop was successfully created." }
-        format.json { render :show, status: :created, location: @shop }
-      else
-        format.html { render :new, status: :unprocessable_entity }
-        format.json { render json: @shop.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # PATCH/PUT /shops/1 or /shops/1.json
-  def update
-    respond_to do |format|
-      if @shop.update(shop_params)
-        format.html { redirect_to shop_url(@shop), notice: "Shop was successfully updated." }
-        format.json { render :show, status: :ok, location: @shop }
-      else
-        format.html { render :edit, status: :unprocessable_entity }
-        format.json { render json: @shop.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # DELETE /shops/1 or /shops/1.json
-  def destroy
-    @shop.destroy
-
-    respond_to do |format|
-      format.html { redirect_to shops_url, notice: "Shop was successfully destroyed." }
-      format.json { head :no_content }
-    end
-  end
-
-  private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_shop
-      @shop = Shop.find(params[:id])
-    end
-
-    # Only allow a list of trusted parameters through.
-    def shop_params
-      params.require(:shop).permit(:name, :address, :phone_number, :duel_space_available, :opening_hours)
-    end
 end
