+++ app/views/admin/shops/_form.html.erb
@@ -1,0 +1,42 @@
+<%= form_with(model: shop) do |form| %>
+  <% if shop.errors.any? %>
+    <div style="color: red">
+      <h2><%= pluralize(shop.errors.count, "error") %> prohibited this shop from being saved:</h2>
+
+      <ul>
+        <% shop.errors.each do |error| %>
+          <li><%= error.full_message %></li>
+        <% end %>
+      </ul>
+    </div>
+  <% end %>
+
+  <div>
+    <%= form.label :name, style: "display: block" %>
+    <%= form.text_field :name %>
+  </div>
+
+  <div>
+    <%= form.label :address, style: "display: block" %>
+    <%= form.text_field :address %>
+  </div>
+
+  <div>
+    <%= form.label :phone_number, style: "display: block" %>
+    <%= form.number_field :phone_number %>
+  </div>
+
+  <div>
+    <%= form.label :duel_space_available, style: "display: block" %>
+    <%= form.number_field :duel_space_available %>
+  </div>
+
+  <div>
+    <%= form.label :opening_hours, style: "display: block" %>
+    <%= form.text_field :opening_hours %>
+  </div>
+
+  <div>
+    <%= form.submit %>
+  </div>
+<% end %>
