+++ app/views/admin/shops/_shop.html.erb
@@ -1,0 +1,27 @@
+<div id="<%= dom_id shop %>">
+  <p>
+    <strong>Name:</strong>
+    <%= shop.name %>
+  </p>
+
+  <p>
+    <strong>Address:</strong>
+    <%= shop.address %>
+  </p>
+
+  <p>
+    <strong>Phone number:</strong>
+    <%= shop.phone_number %>
+  </p>
+
+  <p>
+    <strong>Duel space available:</strong>
+    <%= shop.duel_space_available %>
+  </p>
+
+  <p>
+    <strong>Opening hours:</strong>
+    <%= shop.opening_hours %>
+  </p>
+
+</div>
