+++ app/views/admin/shops/edit.html.erb
@@ -1,0 +1,10 @@
+<h1>Editing shop</h1>
+
+<%= render "form", shop: @shop %>
+
+<br>
+
+<div>
+  <%= link_to "Show this shop", admin_shop_path(@shop) %> |
+  <%= link_to "Back to shops", admin_shops_path %>
+</div>
