+++ app/views/admin/shops/index.html.erb
@@ -11,4 +11,4 @@
   <% end %>
 </div>
 
-
+<%= link_to "New shop", new_admin_shop_path %>
