+++ app/views/admin/shops/new.html.erb
@@ -1,0 +1,9 @@
+<h1>New shop</h1>
+
+<%= render "form", shop: @shop %>
+
+<br>
+
+<div>
+  <%= link_to "Back to shops", admin_shops_path %>
+</div>
