+++ app/views/admin/shops/show.html.erb
@@ -1,0 +1,10 @@
+<p style="color: green"><%= notice %></p>
+
+<%= render @shop %>
+
+<div>
+  <%= link_to "Edit this shop", edit_admin_shop_path(@shop) %> |
+  <%= link_to "Back to shops", admin_shops_path %>
+
+  <%= button_to "Destroy this shop", @shop, method: :delete %>
+</div>
