+++ config/routes.rb
@@ -1,3 +1,9 @@
 Rails.application.routes.draw do
-  resources :shops
+  # 一般ユーザーがアクセスできるショップの一覧と詳細
+  resources :shops, only: [:index, :show]
+
+  # 管理画面用のルーティング
+  namespace :admin do
+    resources :shops
+  end
 end
